feat: add navigation screens definition
Defines a sealed class `NavigationScreens` to represent all possible
navigation destinations in the app.

This class includes top-level destinations like Map, Chat, and Profile,
as well as other screens like Sign In and Add Profile.
Each screen is defined with a route, a user-facing name, and a flag for
htop-level destinations.