feat: add creator to events, unique ID generation
Adds a `creator` field to the `Event` data model to track who created
an event.
The fake event data and associated tests have been updated
to include this new creator information.
Introduces a `getNewID()` function in the `EventRepository` interface
to allow for the generation of unique event IDs. This is implemented
in `FakeEventRepository` using `UUID.randomUUID()`.