feat: add scroll test for tags in user profile
Adds a test tag to the `FlowRow` component that displays user tags.
This allows testing that the tag list is scrollable when it contains too many tags to fit on the screen.

The test `tooManyTagsShowsPartialContent` is renamed to `tooManyTagsImpliesScrollable` and now asserts that the tag list has a scroll action.