refactor: use viewModelScope for coroutines
Moves the coroutine launching for `saveTags` and `loadTags` from the
Composable and `runTest` blocks into the `SelectTagViewModel` using
`viewModelScope`. This improves the separation of concerns by placing
business logic within the ViewModel and making the UI layer responsible
only for triggering actions.

The `SelectTagViewModelTest` is updated to use a`StandardTestDispatcher`
to manage and control the execution of these coroutines during tests.
`advanceUntilIdle()` is used to ensure that all launched coroutines
complete before assertions are made.