refactor: Add Robolectric support to FirebaseAuthUserTest
The `FirebaseAuthUserTest` class is updated to support both local JVM (Robolectric) and Android Instrumented tests.

A constructor parameter `isRobolectric` is introduced to determine the correct host for the Firebase Auth emulator. This avoids hardcoding the host and allows the test base class to be used in different testing environments. The `connect` method in `FirebaseEmulator` is also updated to accept this flag.