refactor: navigation to prevent blocking UI thread
Makes the `resolveUserDestinationScreen` function a `suspend` function
to avoid blocking the main thread while checking for a user profile.

This change introduces a loading state in `MainActivity` to handle the
asynchronous determination of the start destination, showing a
`LinearProgressIndicator` until the navigation route is resolved. The
navigation tests are updated to use `runTest` to accommodate the suspend
function.