feat: Add interactable state to map
Adds a new state `isMapInteractive` to the `MapUiState` to signal when the map is ready for user interaction.

The `MapScreen` now uses this state to apply a test tag, `INTERACTABLE`, which is useful for UI testing. The state is updated to `true` once the TomTom map is fully loaded and initialized.