refactor: move tag observation to LaunchedEffect
Moves the observation of the `tagRepository` from the `init` block of the `SelectTagViewModel` to a new `eventTagRepositoryObserving()` function.

This function is now called from a `LaunchedEffect` in the `SelectTag` composable to ensure the observation coroutine is tied to the component's lifecycle. The view model's constructor has also been simplified by removing the `selectTagMode` parameter.