feat: Cache map marker images to improve performance
Introduce a `MarkerImageCache` to prevent recreating marker images on each recomposition. This singleton object pre-loads and caches all marker pin drawables into `Image` objects in a thread-safe manner.

The `MapScreen` is updated to use this cache for retrieving marker icons, reducing redundant `ImageFactory.fromResource` calls and improving map rendering performance.

Change mapView assignment in the TomTomComposable, it doesn't need to be in the update part of the AndroidView