feat: add event edition flow
Introduce the capability for users to edit existing events. This is achieved by adding a new navigation flow for event editing.
Add `onEditButtonClick` handlers to navigate to the new event edition screen from the map and "My Events" list.
Define new navigation routes and screens: `EventEdition` and `SelectTagEventEdition`.
Create the composable for the `EventEditionScreen`, reusing `EventCreationScreen` with parameters for the event UID and location.
Add new test tags for the event edition screens.