feat: Add ability to change event location from creation form
Add a new `MapChangeLocation` composable to the navigation graph.
Introduce a `onSelectLocation` callback to the `StandardEventCreationForm`.
When the user selects to change the location in the event creation/edition form, switch the `MapViewModel` to `CHANGE_LOCATION` mode and navigate to the map screen.
Add an `onBackChangeLocation` callback to the `MapScreen` to handle navigating back from this new mode.